+++ app/src/pages/AbstractStreamEditPage/StorageNodes/StorageNode/StorageNode.stories.jsx
