+++ app/src/pages/AbstractStreamEditPage/StorageNodes/StorageNode/index.jsx
