+++ app/src/userpages/components/StreamPage/Edit/HistoryView.jsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import Text from '$ui/Text'
 import Select from '$ui/Select'
 import Label from '$ui/Label'
-import Storage from '../shared/Storage'
+import Storage from '$app/src/pages/AbstractStreamEditPage/StorageNodes/Storage'
 
 export const convertFromStorageDays = (days) => {
     let amount = days
